Fix Azure Pipelines Test Runs not getting completed (#1921)
